+++ src/Products/PloneMeeting/tests/testWorkflows.py
@@ -834,7 +834,7 @@ class testWorkflows(PloneMeetingTestCase):
                             pm_logger.info('test_pm_MeetingReviewersValuesAreCorrect: '
                                            'state {0} not found in wf {1}'.format(state, wf.getId()))
 
-    def test_pm_RequiredDataToTriggerTransition(self):
+    def test_pm_RequiredDataToPresentItemCategoryOrGroupsInCharge(self):
         """When MeetingItem.category or MeetingItem.groupsInCharge is used,
            it is required to trigger a transition."""
         cfg = self.meetingConfig
@@ -853,6 +853,32 @@ class testWorkflows(PloneMeetingTestCase):
         self.assertFalse(self.transitions(item))
         item.setCategory('development')
         self.assertTrue(self.transitions(item))
+
+    def test_pm_RequiredDataToPresentItemClassifier(self):
+        """When used, classifier must be set on an item so it may be presented."""
+        self.developers.groups_in_charge = (self.vendors_uid, )
+        self._enableField("classifier")
+        self.changeUser('pmManager')
+        self.create('Meeting')
+        item = self.create('MeetingItem')
+        self.validateItem(item)
+        self.assertEqual(item.wfConditions().mayPresent().msg, u'required_classifier_ko')
+        item.setClassifier('classifier1')
+        self.assertTrue(item.wfConditions().mayPresent())
+
+    def test_pm_RequiredDataToPresentItemProposingGroupWithGroupInCharge(self):
+        """When using proposingGroupWithGroupInCharge, groupsInCharge
+           must be set on an item so it may be presented."""
+        self.developers.groups_in_charge = (self.vendors_uid, )
+        self._enableField("proposingGroupWithGroupInCharge")
+        self.changeUser('pmManager')
+        self.create('Meeting')
+        item = self.create('MeetingItem')
+        self.validateItem(item)
+        self.assertEqual(item.wfConditions().mayPresent().msg, u'required_groupsInCharge_ko')
+        item.setProposingGroupWithGroupInCharge('{0}__groupincharge__{1}'.format(
+            self.developers_uid, self.vendors_uid))
+        self.assertTrue(item.wfConditions().mayPresent())
 
 
 def test_suite():